pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "signed-sealed-delivered"
version = "0.1.0"
description = "Stylometric identity service — extract writing style, transfer it, and embed verifiable signatures"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "spacy>=3.7",
    "numpy>=1.26",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "base58>=2.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ssd"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
